requirements.txt
Django==1.5.1
PyGithub==1.14.2
Pygments==1.6
South==0.7.6
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
awscli==1.2.9
bcdoc==0.12.0
beautifulsoup4==4.1.3
billiard==2.7.3.13
bootstrap==0.1.0
botocore==0.29.0
celery==3.0.9
closure-linter==2.3.13
colorama==0.2.5
creole==1.2
dj-database-url==0.2.1
django-appconf==0.6
django-celery==3.0.9
django-compressor==1.2
django-gravatar==0.1.0
django-redis-cache==0.9.7
docutils==0.11
futures==2.1.4
gunicorn==0.14.6
hiredis==0.1.1
html2text==3.200.3
httpie==0.8.0
ipython==0.13.2
jmespath==0.2.1
kombu==2.4.6
newrelic==1.13.1.31
numpy==1.7.1
ply==3.4
psycopg2==2.4.6
pyOpenSSL==0.13.1
pyasn1==0.1.7
pydot==1.0.2
pyflakes==0.8.1
pyparsing==1.5.7
python-dateutil==2.2
python-gflags==2.0
pytz==2012d
raven==3.3.3
redis==2.7.2
requests==2.2.1
rsa==3.1.2
six==1.3.0
virtualenv==1.11.6
wsgiref==0.1.2
